+++ src/NuGet.Updater.Tests/Entities/TestPackageSource.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,18 +10,26 @@ namespace NuGet.Updater.Tests.Entities
 {
 	public class TestPackageSource : IUpdaterSource
 	{
-		public TestPackageSource(Uri uri, params TestPackage[] packages)
+		private readonly Dictionary<string, string[]> _packages;
+
+		public TestPackageSource(Uri uri, Dictionary<string, string[]> packages)
 		{
 			Uri = uri;
-			//Packages = packages?.Select(p => new NuGetPackage(p, Uri)).ToArray();
+			_packages = packages;
 		}
 
 		public Uri Uri { get; }
 
-		public UpdaterPackage[] Packages { get; }
+		public async Task<UpdaterPackage> GetPackage(CancellationToken ct, PackageReference reference, Logger log = null)
+		{
+			var packageId = reference.Id;
 
-		public async Task<UpdaterPackage> GetPackage(CancellationToken ct, PackageReference reference, Logger log = null) => Packages.FirstOrDefault(p => p.PackageId == reference.Id);
+			var versions = _packages
+				.GetValueOrDefault(reference.Id)
+				?.Select(v => new UpdaterVersion(v, Uri))
+				.ToArray() ?? new UpdaterVersion[0];
 
-		public async Task<UpdaterPackage[]> GetPackages(CancellationToken ct, Logger log = null) => Packages;
+			return new UpdaterPackage(reference, versions);
+		}
 	}
 }