+++ src/NuGet.Updater.Tests/UpdaterPackagePackageTests.cs
@@ -1,16 +1,16 @@
 ﻿using System;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NuGet.Updater.Entities;
 using NuGet.Updater.Extensions;
-using NuGet.Updater.Tests.Entities;
 
 namespace NuGet.Updater.Tests
 {
 	[TestClass]
 	public class UpdaterPackagePackageTests
 	{
+		private static readonly Uri TestFeedUri = new Uri("http://localhost");
+
 		[TestMethod]
 		public async Task GivenPackageWithMatchingVersion_VersionIsFound()
 		{
@@ -20,7 +20,7 @@ namespace NuGet.Updater.Tests
 			};
 
 			var packageVersion = "1.0-beta.1";
-			var package = new UpdaterPackage(new TestPackage("nventive.NuGet.Updater", packageVersion), new Uri("http://localhost"));
+			var package = new UpdaterPackage("nventive.NuGet.Updater", TestFeedUri, packageVersion);
 
 			var version = package.GetLatestVersion(parameters);
 
@@ -36,7 +36,7 @@ namespace NuGet.Updater.Tests
 				TargetVersions = new[] { "stable" },
 			};
 
-			var package = new UpdaterPackage(new TestPackage("nventive.NuGet.Updater", "1.0-beta.1"), new Uri("http://localhost"));
+			var package = new UpdaterPackage("nventive.NuGet.Updater", TestFeedUri, "1.0-beta.1");
 
 			var version = package.GetLatestVersion(parameters);