+++ src/NuGet.Updater/Entities/IUpdaterSource.cs
@@ -6,8 +6,6 @@ namespace NuGet.Updater.Entities
 {
 	public interface IUpdaterSource
 	{
-		Task<UpdaterPackage[]> GetPackages(CancellationToken ct, Logger log = null);
-
 		Task<UpdaterPackage> GetPackage(CancellationToken ct, PackageReference reference, Logger log = null);
 	}
 }