+++ src/NuGet.Updater/Entities/UpdaterVersion.cs
@@ -4,16 +4,21 @@ using NuGet.Versioning;
 namespace NuGet.Updater.Entities
 {
 	public class UpdaterVersion : IComparable<UpdaterVersion>
-	{ 
-		public UpdaterVersion(Uri feedUri, NuGetVersion version)
+	{
+		internal UpdaterVersion(string version, Uri feedUri)
+			: this(new NuGetVersion(version), feedUri)
 		{
-			FeedUri = feedUri;
-			Version = version;
 		}
 
-		public Uri FeedUri { get; }
+		public UpdaterVersion(NuGetVersion version, Uri feedUri)
+		{
+			Version = version;
+			FeedUri = feedUri;
+		}
 
 		public NuGetVersion Version { get; }
+
+		public Uri FeedUri { get; }
 
 		public int CompareTo(UpdaterVersion other) => Version.CompareTo(other.Version);
 	}