+++ src/NuGet.Updater/Helpers/SolutionHelper.cs
@@ -6,39 +6,27 @@ using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Updater.Entities;
 using NuGet.Updater.Extensions;
+using NuGet.Updater.Log;
 using Uno.Extensions;
 
 namespace NuGet.Updater.Helpers
 {
 	public static class SolutionHelper
 	{
-		internal static async Task<string[]> GetTargetFilePaths(CancellationToken ct, string solutionPath, UpdateTarget updateTarget)
+		internal static async Task<PackageReference[]> GetPackageReferences(
+			CancellationToken ct,
+			string solutionPath,
+			UpdateTarget updateTarget,
+			Logger log = null
+		)
 		{
-			switch(updateTarget)
-			{
-				case UpdateTarget.Csproj:
-					return await GetProjectFiles(ct, solutionPath);
-				case UpdateTarget.DirectoryProps:
-				case UpdateTarget.DirectoryTargets:
-					return new[] { GetDirectoryFile(solutionPath, updateTarget) };
-				case UpdateTarget.Nuspec:
-					return new string[0]; // find all matching files in the solution folder
-				case UpdateTarget.ProjectJson:
-					return new string[0]; //find all csproj, look for project.json next to them
-				case UpdateTarget.Unspecified:
-				case UpdateTarget.All:
-				default:
-					return new string[0];
-			}
-		}
+			log?.Write($"Retrieving references from files in {solutionPath}");
 
-		internal static async Task<PackageReference[]> GetPackageReferences(CancellationToken ct, string solutionPath, UpdateTarget updateTarget)
-		{
 			var packages = new List<PackageReference>();
 
 			if(updateTarget.Matches(UpdateTarget.Csproj))
 			{
-				foreach(var f in await GetProjectFiles(ct, solutionPath))
+				foreach(var f in await GetProjectFiles(ct, solutionPath, log))
 				{
 					packages.AddRange(await GetFileReferences(ct, f, UpdateTarget.Csproj));
 				}
@@ -46,44 +34,49 @@ namespace NuGet.Updater.Helpers
 
 			if(updateTarget.Matches(UpdateTarget.DirectoryProps))
 			{
-				packages.AddRange(await GetFileReferences(ct, GetDirectoryFile(solutionPath, UpdateTarget.DirectoryProps), UpdateTarget.DirectoryProps));
+				packages.AddRange(await GetFileReferences(ct, GetDirectoryFile(solutionPath, UpdateTarget.DirectoryProps, log), UpdateTarget.DirectoryProps));
 			}
 
 			if(updateTarget.Matches(UpdateTarget.DirectoryTargets))
 			{
-				packages.AddRange(await GetFileReferences(ct, GetDirectoryFile(solutionPath, UpdateTarget.DirectoryTargets), UpdateTarget.DirectoryTargets));
+				packages.AddRange(await GetFileReferences(ct, GetDirectoryFile(solutionPath, UpdateTarget.DirectoryTargets, log), UpdateTarget.DirectoryTargets));
 			}
 
 			return packages
 				.GroupBy(p => p.Id)
 				.Select(g => new PackageReference(
 					g.Key,
-					g.Select(p => p.Version).OrderBy(v => v).FirstOrDefault(),
+					g.Select(p => p.Version).OrderByDescending(v => v).FirstOrDefault(),
 					g.SelectMany(p => p.Files).GroupBy(f => f.Key).ToDictionary(f => f.Key, f => f.SelectMany(x => x.Value).Distinct().ToArray())
 				))
 				.ToArray();
 		}
 
-		private static async Task<string[]> GetProjectFiles(CancellationToken ct, string solutionPath)
+		private static async Task<string[]> GetProjectFiles(CancellationToken ct, string solutionPath, Logger log = null)
 		{
 			var solutionContent = await FileHelper.ReadFileContent(ct, solutionPath);
 			var solutionFolder = Path.GetDirectoryName(solutionPath);
 
 			var matches = Regex.Matches(solutionContent, "[^\\s\"]*\\.csproj");
 
-			return matches
+			var files = matches
 				.Cast<Match>()
 				.Select(m => Path.Combine(solutionFolder, m.Value))
 				.ToArray();
+
+			log?.Write($"Found {files.Length} csproj files");
+
+			return files;
 		}
 
-		private static string GetDirectoryFile(string solutionPath, UpdateTarget target)
+		private static string GetDirectoryFile(string solutionPath, UpdateTarget target, Logger log = null)
 		{
 			var solutionFolder = Path.GetDirectoryName(solutionPath);
 			var file = Path.Combine(solutionFolder, target.GetDescription());
 
 			if(File.Exists(file))
 			{
+				log?.Write($"Found {target.GetDescription()}");
 				return file;
 			}